refactor 참가자의 수validation 위치 이동
