+++ src/main/java/nextstep/ladder/controller/LadderController.java
@@ -17,14 +17,14 @@ public final class LadderController {
 
     public void run() {
         List<String> participantNames = InputTable.insertParticipantNames();
-        Ladder ladder = createLadder(InputTable.ladderHeight(), gameParticipantCount(participantNames));
+        Ladder ladder = createLadder(gameParticipantCount(participantNames), InputTable.ladderHeight());
     }
 
     public int gameParticipantCount(List<String> participantsNames) {
         return ladderService.gameParticipantCount(participantsNames);
     }
 
-    public Ladder createLadder(int height, int participant) {
-        return ladderService.createLadder(height, participant);
+    public Ladder createLadder(int participant, int height) {
+        return ladderService.createLadder(participant, height);
     }
 }
