+++ src/test/java/nextstep/ladder/model/ParticipantsTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class ParticipantsTest {
     @Test
@@ -14,5 +15,14 @@ class ParticipantsTest {
         List<String> names = List.of("a", "b", "c");
         Participants participants = new Participants(names);
         assertThat(names).hasSize(participants.size());
+    }
+
+    @Test
+    @DisplayName("참가자는 최소 2명이상이어야 합니다.")
+    void createTwoBars() {
+        List<String> names = List.of("a");
+        assertThrows(IllegalArgumentException.class, () -> {
+            new Participants(names);
+        });
     }
 }
